refactor: lodash imports
