+++ src/components/shared/index.ts
@@ -1,6 +1,6 @@
 export * from './Caret';
 export * from './Description';
-export * from './Description';
+export * from './Divider';
 export * from './Property';
 export * from './Types';
 export * from './Validations';
